complete remap of folder system, flattened it
REMAP: Originally the file system was nickswebsite/ git files, readme, and a
single folder containing the actual django app. now the top level files
of that django app folder: blog, blogdocs, nickswebsite, all now live in
the top level with the git files and the readme. makes for a much better
flat file system. this avoids the nickswebsite/nickswebsite/nickswebsite
issues that I used to have when deploying! renamed the the repo
nicksblog, and left the django website as nickswebsite and the blog app
as blog.

ADDED: Bower installed bootstrap, jquery, but have not yet reimplimented
the design.